- fixed Exceptions when deleting not existing record from db in DriversManager/VehiclesManager - added classes: Statement / StementsController / StatementsManager - added mapping file for Statement - added mapping resource in hibernate.cfg  for Statement class - changed relationship between drivers and coordinates from primary_key to foregin_key

+++ src/main/java/skip/DriversManager.java
@@ -112,7 +112,8 @@ public class DriversManager {
 		Session session = HibernateUtil.getSessionFactory().getCurrentSession();
 		session.beginTransaction();
 		Driver d = (Driver) session.get(Driver.class, id);
-		session.delete(d);
+                if(d != null)
+                    session.delete(d);
 		session.getTransaction().commit();
 		return d;
 	}
